fix bug: time wasting clock not fully restored
When the value of the time wasting clock was restored from the browser's
local storage, the hidden form field in the end of class form was left
with a value of zero.  So if the lesson ends after a browser restart
with no further time wasted, the time on the clock was not submitted.